requirements.txt
requests
python-dotenv
backoff 
urllib3==1.26.15